pyproject.toml
[tool.poetry]
name = "ib_value_averaging"
version = "0.1.0"
description = ""
authors = ["Matthew Cross <mattyc246@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
ibapi = "^9.76.1"
python-dotenv = "^0.14.0"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
